requirements.txt
# Core
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.4.0
matplotlib==3.9.2
tqdm==4.66.5

# ML/DL
torch --pre --index-url https://download.pytorch.org/whl/nightly/cpu
torchvision --pre --index-url https://download.pytorch.org/whl/nightly/cpu
torchaudio --pre --index-url https://download.pytorch.org/whl/nightly/cpu

# Extras
torchcam==0.4.0
reportlab==4.4.3
jinja2==3.1.6
